Extract `ServerSettingsWriter`
