+++ app/core/src/main/java/com/fsck/k9/preferences/AccountSettingsWriter.kt
@@ -8,9 +8,6 @@ import com.fsck.k9.AccountPreferenceSerializer.Companion.OUTGOING_SERVER_SETTING
 import com.fsck.k9.Core
 import com.fsck.k9.Preferences
 import com.fsck.k9.ServerSettingsSerializer
-import com.fsck.k9.mail.AuthType
-import com.fsck.k9.mail.ConnectionSecurity
-import com.fsck.k9.mail.ServerSettings
 import com.fsck.k9.mailstore.SpecialLocalFoldersCreator
 import java.util.UUID
 import kotlinx.datetime.Clock
@@ -19,11 +16,12 @@ internal class AccountSettingsWriter(
     private val preferences: Preferences,
     private val localFoldersCreator: SpecialLocalFoldersCreator,
     private val clock: Clock,
-    private val serverSettingsSerializer: ServerSettingsSerializer,
+    serverSettingsSerializer: ServerSettingsSerializer,
     private val context: Context,
 ) {
     private val identitySettingsWriter = IdentitySettingsWriter()
     private val folderSettingsWriter = FolderSettingsWriter()
+    private val serverSettingsWriter = ServerSettingsWriter(serverSettingsSerializer)
 
     fun write(account: ValidatedSettings.Account): Pair<AccountDescription, AccountDescription> {
         val editor = preferences.createStorageEditor()
@@ -57,8 +55,16 @@ internal class AccountSettingsWriter(
             value = messageNotificationChannelVersion,
         )
 
-        writeServerSettings(editor, key = "$accountUuid.$INCOMING_SERVER_SETTINGS_KEY", server = account.incoming)
-        writeServerSettings(editor, key = "$accountUuid.$OUTGOING_SERVER_SETTINGS_KEY", server = account.outgoing)
+        serverSettingsWriter.writeServerSettings(
+            editor,
+            key = "$accountUuid.$INCOMING_SERVER_SETTINGS_KEY",
+            server = account.incoming,
+        )
+        serverSettingsWriter.writeServerSettings(
+            editor,
+            key = "$accountUuid.$OUTGOING_SERVER_SETTINGS_KEY",
+            server = account.outgoing,
+        )
 
         writeIdentities(editor, accountUuid, account.identities)
         writeFolders(editor, accountUuid, account.folders)
@@ -136,48 +142,5 @@ internal class AccountSettingsWriter(
 
     private fun isAccountNameUsed(name: String?, accounts: List<Account>): Boolean {
         return accounts.any { it.displayName == name }
-    }
-
-    private fun writeServerSettings(
-        editor: StorageEditor,
-        key: String,
-        server: ValidatedSettings.Server,
-    ) {
-        val serverSettings = createServerSettings(server)
-        val serverSettingsJson = serverSettingsSerializer.serialize(serverSettings)
-        editor.putStringWithLogging(key, serverSettingsJson)
-    }
-
-    private fun createServerSettings(server: ValidatedSettings.Server): ServerSettings {
-        val connectionSecurity = convertConnectionSecurity(server.connectionSecurity)
-        val authenticationType = AuthType.valueOf(server.authenticationType)
-        val password = if (authenticationType == AuthType.XOAUTH2) "" else server.password
-
-        return ServerSettings(
-            server.type,
-            server.host,
-            server.port,
-            connectionSecurity,
-            authenticationType,
-            server.username,
-            password,
-            server.clientCertificateAlias,
-            server.extras,
-        )
-    }
-
-    @Suppress("TooGenericExceptionCaught", "SwallowedException")
-    private fun convertConnectionSecurity(connectionSecurity: String): ConnectionSecurity {
-        return try {
-            // TODO: Add proper settings validation and upgrade capability for server settings. Once that exists, move
-            //  this code into a SettingsUpgrader.
-            when (connectionSecurity) {
-                "SSL_TLS_OPTIONAL" -> ConnectionSecurity.SSL_TLS_REQUIRED
-                "STARTTLS_OPTIONAL" -> ConnectionSecurity.STARTTLS_REQUIRED
-                else -> ConnectionSecurity.valueOf(connectionSecurity)
-            }
-        } catch (e: Exception) {
-            ConnectionSecurity.SSL_TLS_REQUIRED
-        }
     }
 }
